Pequeña corrección en UserRepository y agregué que se muestren imagenes por su descripción corta en productos destacados

+++ ecommercesportsDEV/src/main/java/com/ecommercesports/ecommercesports/controllers/UserController.java
@@ -230,7 +230,7 @@ public class UserController {
     }
 
 	@PostMapping("/delete")
-	public RedirectView delete(long idUser) {
+	public RedirectView delete(int idUser) {
 		userRoleRepository.deleteById(idUser);
 		userRepository.deleteById(idUser);
